--- src/App.css
@@ -1,191 +1,0 @@
-.App {
-  text-align: left;
-  color: #333;
-}
-
-.App-header {
-  padding: 20px;
-  padding-top: 0;
-  padding-bottom: 0;
-  color: white;
-  background-color: #292929;
-  flex-direction: row;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  /*background-color: #0061ff;*/
-}
-
-.App-header-container {
-  max-width: 800px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  flex: 1;
-}
-
-.App-title {
-  margin: 0;
-  font-size: 1.3em;
-  font-weight: 700;
-}
-
-.App-title-link {
-  color: white;
-  text-decoration: none;
-  font-weight: 700;
-}
-
-
-.App-title-link:visited {
-  color: white;
-  text-decoration: none;
-}
-
-.App-title-link:hover {
-  color: white;
-  text-decoration: none;
-}
-
-.App-title-link:active {
-  color: white;
-  text-decoration: none;
-}
-
-.App-intro {
-  font-size: normal;
-}
-
-.App-body {
-  padding: 10px;
-}
-
-.App-window {
-  background-color: #fff;
-  color: #182535;
-  padding: 10px;
-  width: 100%;
-  max-width: 800px;
-  margin: 10px;
-  /*border-radius: 2px;
-  border-style: solid;
-  border-width: 1px;
-  border-color: #ddd;*/
-}
-
-.App-container {
-  color: black;
-  background-color: #fff;
-  /*padding: 20px;
-  border-radius: 2px;
-  margin-bottom: 10px;
-  border-style: solid;
-  border-width: 1px;
-  border-color: #ddd;*/
-  display: flex;
-  flex-direction: column;
-}
-
-.App-space {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.header-bold {
-  display: flex;
-  font-weight: 400;
-}
-
-.container-header {
-  font-family: monospace;
-  font-size: 1.25em;
-  color: #aaa;
-  font-weight: 400;
-  border-bottom-style: dotted;
-  border-bottom-width: 1px;
-  border-bottom-color: black;
-}
-
-.container-header-span {
-  margin-bottom: 10px;
-}
-
-.window-header {
-  font-size: 1.3em;
-  display: flex;
-  flex-direction: row;
-  border-bottom-style: solid;
-  border-bottom-width: 2px;
-  border-bottom-color: #0061ff;
-  margin-bottom: 20px;
-}
-
-.window-header-selected {
-  font-weight: 700;
-  padding-bottom: 10px;
-}
-
-.window-header-bar {
-  padding-left: 10px;
-  padding-right: 10px;
-  font-weight: 300;
-  color: #0061ff;
-}
-
-.window-header-deselected {
-  font-weight: 400;
-  cursor: pointer;
-}
-
-.window-header-deselected:hover {
-  text-decoration: underline;
-}
-
-.App-search-input {
-  background-color: #555;
-  color: white;
-  border-width: 1px;
-  border-color: #222;
-  max-width: 400px;
-}
-
-.App-search-button {
-  background-color: #0061ff;
-  color: white;
-  border: none;
-  cursor: pointer;
-}
-
-input[type=submit] {
-  background-color: #0061ff;
-  color: white;
-  border: none;
-  cursor: pointer;
-}
-
-label {
-  font-weight: bold;
-  font-size: 0.9em;
-}
-
-input {
-  border-radius: 2px;
-  padding: 10px;
-  font-size: 1.1em;
-  font-family: monospace;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  border-width: 1px;
-  border-color: #222;
-}
-.container-textarea {
-  margin-top: 20px;
-  margin-bottom: 20px;
-  border-width: 1px;
-  border-color: #222;
-  font-size: 1.1em;
-  padding: 10px;
-  border-radius: 2px;
-}
