+++ src/components/Window.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import '../App.css'
+import './App.css'
 
 class Window extends Component {
   constructor () {
